+++ src/comp/c_esp.c
@@ -71,9 +71,9 @@ static bool c_esp_check_context(const struct c_context *context,
 
 static int c_esp_encode(struct c_context *const context,
                         const struct ip_packet *ip,
-                        const int packet_size,
+                        const size_t packet_size,
                         unsigned char *const dest,
-                        const int dest_size,
+                        const size_t dest_size,
                         rohc_packet_t *const packet_type,
                         int *const payload_offset);
 
@@ -486,9 +486,9 @@ bad_context:
  */
 static int c_esp_encode(struct c_context *const context,
                         const struct ip_packet *ip,
-                        const int packet_size,
+                        const size_t packet_size,
                         unsigned char *const dest,
-                        const int dest_size,
+                        const size_t dest_size,
                         rohc_packet_t *const packet_type,
                         int *const payload_offset)
 {
